separate core/v1beta1/helper tests and rename  managedresource
