+++ pkg/apis/resources/v1alpha1/helper/managedresource_test.go
